LayoutTests/crypto/subtle/rsassa-pkcs1-v1_5-import-key-sign-sha256-expected.txt

Test signing with RSASSA-PKCS1-v1_5 SHA-256 using an imported key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bytesToHexString(signature) is expectedSignature
PASS successfullyParsed is true

TEST COMPLETE
